requirements.txt
absl-py==1.4.0
altgraph==0.17.3
async-timeout==4.0.2
attrs==22.2.0
Automat==22.10.0
certifi==2022.12.7
charset-normalizer==3.1.0
constantly==15.1.0
contourpy==1.0.7
cycler==0.11.0
dlib-binary==19.21.99
flatbuffers==23.3.3
fonttools==4.39.2
hyperlink==21.0.0
idna==3.4
ifaddr==0.2.0
importlib-resources==5.12.0
incremental==22.10.0
joblib==1.2.0
kiwisolver==1.4.4
klein==21.8.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mediapipe==0.9.1.0
numpy==1.24.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
packaging==23.0
Pillow==9.4.0
protobuf==3.20.3
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
tomli==2.0.1
Tubes==0.2.1
Twisted==22.10.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
zeroconf==0.47.3
zipp==3.15.0
zope.interface==6.0
